Major overhaul of the build system
This is a large change, but since it involved the changes to the build
system, it is extremely hard to split into smaller chunks.

Use automake to generate makefiles. This documentation was used to write
build rules for python wrappers in Makefile.am:
https://github.com/swig/swig/wiki/FAQ#autotools-integration

configure.ac rewritten substantially to use existing checks and
canonical m4sh constructs, where applicable.

The check for LibFTDI version in the source code has been changed to a
canonical "HAVE_LIBFTDI".